✨  F12: BASKET > Enable live edit by clicking on basket products
