+++ src/components/pages/Order/Main/Basket/BasketProducts.jsx
@@ -7,17 +7,21 @@ import { useContext } from "react";
 import OrderContext from "../../../../../context/OrderContext.jsx";
 
 export default function BasketProducts() {
-  const { basket, handleDeleteBasketProduct, isModeAdmin, menu } = useContext(OrderContext);
+  const { basket, handleDeleteBasketProduct, isModeAdmin, menu, handleProductSelected } = useContext(OrderContext);
 
-  const handleOnDelete = (id) => {
+  const handleOnDelete = (event,id) => {
+    event.stopPropagation();
     handleDeleteBasketProduct(id);
   };
+
+
 
   return (
     <BasketProductsStyled>
       {basket.map((basketProduct) => {
         const menuProduct = findObjectById(basketProduct.id, menu)
         // console.log("menuProduct: ", menuProduct);
+
 
         return (
           <div className="basket-card" key={basketProduct.id}>
@@ -29,9 +33,10 @@ export default function BasketProducts() {
                   : IMAGE_COMMING_SOON
               }
               quantity={basketProduct.quantity}
-              onDelete={() => handleOnDelete(basketProduct.id)}
+              onDelete={(event) => handleOnDelete(event, basketProduct.id)}
               isClickable={isModeAdmin}
-            />
+              onClick={isModeAdmin ? () => handleProductSelected(basketProduct.id) : null}
+              />
           </div>
         );
       })}
