+++ src/components/pages/Order/Main/Basket/Total.jsx
@@ -4,17 +4,19 @@ import { theme } from "../../../../..";
 import Header from "../../../../reusable-ui/Header";
 import { useContext } from "react";
 import OrderContext from "../../../../../context/OrderContext";
-import { calculateSumToPay, formatPrice } from "../../../../../utils/maths";
+import { formatPrice } from "../../../../../utils/maths";
+import { findObjectById } from "../../../../../utils/array";
+import { calculateSumToPay } from "./helper";
 
 export default function Total() {
   const { basket, menu } = useContext(OrderContext);
 
-  const sumToPay = calculateSumToPay(basket, menu);
+  const sumToPay = calculateSumToPay(basket, menu)
 
   return (
     <Header>
       <TotalStyled>
-        <span className="toal">Total</span>
+        <span className="total">Total</span>
         <span className="amount">{formatPrice(sumToPay)}</span>
       </TotalStyled>
     </Header>
@@ -32,3 +34,4 @@ const TotalStyled = styled.div`
   font-weight: ${theme.fonts.weights.bold};
   letter-spacing: 2px;
 `;
+
