+++ src/utils/maths.jsx
@@ -1,5 +1,3 @@
-import { findObjectById } from "./array"
-
 export function formatPrice(priceToFormat) {
   let price = priceToFormat
 
@@ -20,11 +18,3 @@ export function replaceFrenchCommaWithDot(price) {
 }
 
 
-export const calculateSumToPay = (basket, menu) => {
-  return basket.reduce((total, basketProduct) => {
-    const menuProduct = findObjectById(basketProduct.id, menu);
-
-    total += menuProduct.price * basketProduct.quantity;
-    return total;
-  }, 0);
-}
